Upgrade to new React alpha version (#568)
* Upgrade to new React alpha version

* Remove try-catch on app-server

* Fix lint

* Revert to previous tsconfig lib

* Re-enable typescript lint in CI

+++ src/components/pages/StatisticsPage.tsx
@@ -435,7 +435,7 @@ const StatisticsPage: React.FC = () => {
                         : daysInterval <= 365
                         ? 'long'
                         : undefined,
-                    year: daysInterval > 365 ? 'long' : undefined,
+                    year: daysInterval > 365 ? 'numeric' : undefined,
                   })
                 }
                 tickSizeOuter={0}
